Fixed TransferRepository ID Type

+++ src/main/java/io/droksty/transfersdemo/repository/TransferRepository.java
@@ -8,7 +8,7 @@ import java.time.LocalDate;
 import java.util.List;
 
 @Repository
-public interface TransferRepository extends JpaRepository<Transfer, Integer> {
+public interface TransferRepository extends JpaRepository<Transfer, Long> {
     List<Transfer> findAllByPickupDateEqualsOrderByPickupTimeAsc(LocalDate date);
     List<Transfer> findAllByPickupDateBetweenOrderByPickupDateAscPickupTimeAsc(LocalDate from, LocalDate to);
